refactor: big refactoring for runtime module
