+++ public/manifest2.json
@@ -5,7 +5,8 @@
     "default_locale": "en",
     "version": "1.11.0",
     "web_accessible_resources": [
-        "icons/*",
+        "images/*",
+        "sounds/*",
         "assets/*"
     ],
     "permissions": [
@@ -13,15 +14,15 @@
         "tabs"
     ],
     "icons": {
-        "32": "icons/icon-32.png",
-        "48": "icons/icon-48.png",
-        "64": "icons/icon-64.png",
-        "128": "icons/icon-128.png"
+        "32": "images/icons/icon-32.png",
+        "48": "images/icons/icon-48.png",
+        "64": "images/icons/icon-64.png",
+        "128": "images/icons/icon-128.png"
     },
     "browser_action": {
         "default_icon": {
-            "32": "icons/icon-32.png",
-            "48": "icons/icon-48.png"
+            "32": "images/icons/icon-32.png",
+            "48": "images/icons/icon-48.png"
         },
         "default_title": "Tab Guardian",
         "default_popup": "index.html"
