+++ src/popup/components/Navbar/Sidebar/Heading.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref, onMounted, computed } from 'vue'
 import { trans } from '@common/modules/trans'
-import { getImageURL } from '@common/modules/browser/runtime'
+import { getImageURL } from '@common/modules/browser/url'
 import { isDevelopment } from '@common/modules/isDevelopment'
 import { isFirefox } from '@common/modules/browser/isFirefox'
 import { getLocalStorageUsage } from '@common/modules/storage/getLocalStorageUsage'
