+++ src/popup/components/Navbar/Sidebar/Sidebar.vue
@@ -3,7 +3,7 @@ import { useSidebarStore } from '@/stores/sidebar'
 import { trans } from '@common/modules/trans'
 import { openSettingsPage } from '@/modules/openSettingsPage'
 import { isFirefox } from '@common/modules/browser/isFirefox'
-import { getImageURL } from '@common/modules/browser/runtime'
+import { getImageURL } from '@common/modules/browser/url'
 import Hamburger from '@/components/Navbar/Sidebar/Hamburger.vue'
 import RightSlideTransition from '@common/components/Transitions/RightSlideTransition.vue'
 import Overlay from '@/components/Navbar/Sidebar/Overlay.vue'
