+++ src/popup/components/Views/GroupView/GroupControls/Actions/StashTabsAction.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { Group } from '@/types'
 import Control from '@/components/Control.vue'
-import downImage from '@/assets/images/tab-icons/down.png'
+import { getImageURL } from '@common/modules/browser/url'
 import { useTabsStore } from '@/stores/tabs'
 import { trans } from '@common/modules/trans'
 
@@ -19,6 +19,10 @@ function stashTabs(): void {
         @click="stashTabs"
         class="bg-slate-300 dark:bg-slate-700"
     >
-        <img :src="downImage" alt="Stash tabs" class="dark:invert" />
+        <img
+            :src="getImageURL('tab-icons/down.png')"
+            alt="Stash tabs"
+            class="dark:invert"
+        />
     </Control>
 </template>
